Defines methods show(), hide(), and toggle()
These methods are used for showing, hiding, and toggling the display of elements on a web page. The methods loop through all the elements in the current selection, set or toggle their display CSS property, and return the current object to enable chaining. If an element doesn't have a style property, it is skipped in the loop. The $ object is not defined in the code snippet and may refer to a custom JavaScript library.